Revert "Merge pull request #30169 from software-mansion-labs/ts-migration/personal-details-utils"
This reverts commit e17c313b0a7c4cedcd75c5ceae3ac577585996a0, reversing
changes made to 9973d8fd4e847e9f635c792b4678f642ced5d7cd.
