+++ src/types/onyx/index.ts
@@ -60,6 +60,7 @@ import WalletTransfer from './WalletTransfer';
 
 export type {
     Account,
+    UserLocation,
     AccountData,
     AddDebitCardForm,
     BankAccount,
@@ -91,16 +92,16 @@ export type {
     PersonalDetailsList,
     PlaidData,
     Policy,
-    PolicyCategories,
     PolicyCategory,
+    PolicyCategories,
     PolicyMember,
     PolicyMembers,
     PolicyTag,
     PolicyTags,
     PrivatePersonalDetails,
-    RecentWaypoint,
     RecentlyUsedCategories,
     RecentlyUsedTags,
+    RecentWaypoint,
     ReimbursementAccount,
     ReimbursementAccountDraft,
     Report,
@@ -119,7 +120,6 @@ export type {
     Transaction,
     TransactionViolation,
     User,
-    UserLocation,
     UserWallet,
     ViolationName,
     WalletAdditionalDetails,
